+++ test/widgets/wishlist_deletion_test.dart
@@ -1,9 +1,11 @@
+import 'package:MobileOne/arguments/arguments.dart';
 import 'package:MobileOne/data/wishlist.dart';
 import 'package:MobileOne/localization/delegate.dart';
 import 'package:MobileOne/localization/supported.dart';
 import 'package:MobileOne/pages/openedListPage.dart';
 import 'package:MobileOne/providers/itemsList_provider.dart';
 import 'package:MobileOne/providers/wishlist_head_provider.dart';
+import 'package:MobileOne/utility/arguments.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -41,6 +43,8 @@ class MockWishlistHeadProvider extends Mock implements WishlistHeadProvider {}
 
 class MockWishlist extends Mock implements Wishlist {}
 
+class MockArguments extends Mock implements Arguments {}
+
 void main() {
   setSupportedLocales([Locale('fr', 'FR')]);
 
@@ -55,6 +59,13 @@ void main() {
     final wishlist = MockWishlist();
     when(wishlist.label).thenReturn("test");
     when(_wishlistHeadProvider.getWishlist(any)).thenReturn(wishlist);
+
+    final mockArguments = MockArguments();
+    Arguments.proxy = mockArguments;
+    when(mockArguments.get(any)).thenReturn(OpenedListArguments(
+      listUuid: "",
+      isGuest: false,
+    ));
 
     await tester.pumpWidget(buildTestableWidget(new OpenedListPage()));
     await tester.pumpAndSettle(Duration(seconds: 2));